+++ src/hittable/medium.rs
@@ -4,11 +4,11 @@ use rand::Rng;
 pub struct ConstantMedium {
     pub hittable: Box<dyn Hittable>,
     pub phase_function: Material,
-    pub neg_inv_density: f64,
+    pub neg_inv_density: f32,
 }
 
 impl ConstantMedium {
-    pub fn new(hittable: Box<dyn Hittable>, density: f64, color: Color) -> Self {
+    pub fn new(hittable: Box<dyn Hittable>, density: f32, color: Color) -> Self {
         Self {
             hittable,
             neg_inv_density: (-1.0 / density),
@@ -18,19 +18,19 @@ impl ConstantMedium {
 }
 
 impl Hittable for ConstantMedium {
-    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64, hit_record: &mut HitRecord) -> bool {
+    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32, hit_record: &mut HitRecord) -> bool {
         let mut rec1 = HitRecord::default();
         let mut rec2 = HitRecord::default();
 
         if !self
             .hittable
-            .hit(ray, -f64::INFINITY, f64::INFINITY, &mut rec1)
+            .hit(ray, -f32::INFINITY, f32::INFINITY, &mut rec1)
         {
             return false;
         }
         if !self
             .hittable
-            .hit(ray, rec1.t + 0.0001, f64::INFINITY, &mut rec2)
+            .hit(ray, rec1.t + 0.0001, f32::INFINITY, &mut rec2)
         {
             return false;
         }
@@ -50,7 +50,7 @@ impl Hittable for ConstantMedium {
 
         let ray_length = ray.direction.length();
         let distance_inside_boundary = (rec2.t - rec1.t) * ray_length;
-        let hit_distance = self.neg_inv_density * rand::thread_rng().gen::<f64>().ln();
+        let hit_distance = self.neg_inv_density * rand::thread_rng().gen::<f32>().ln();
 
         if hit_distance > distance_inside_boundary {
             return false;
@@ -67,7 +67,7 @@ impl Hittable for ConstantMedium {
     }
 
     #[allow(unused_variables)]
-    fn bounding_box(&self, t0: f64, t1: f64) -> Option<AABB> {
+    fn bounding_box(&self, t0: f32, t1: f32) -> Option<AABB> {
         self.hittable.bounding_box(t0, t1)
     }
 }
