+++ src/hittable_list.rs
@@ -22,7 +22,7 @@ impl<'a> HittableList {
 }
 
 impl<'a> Hittable for HittableList {
-    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64, hit_record: &mut HitRecord) -> bool {
+    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32, hit_record: &mut HitRecord) -> bool {
         let mut hit_anything = false;
         let mut closest_so_far = t_max;
 
@@ -38,7 +38,7 @@ impl<'a> Hittable for HittableList {
         return hit_anything;
     }
 
-    fn bounding_box(&self, t0: f64, t1: f64) -> Option<AABB> {
+    fn bounding_box(&self, t0: f32, t1: f32) -> Option<AABB> {
         if self.objects.is_empty() {
             return None;
         }
